pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "samchully"
version = "1.0.0"
description = "Project samchully with smartmind"
authors = ["KyungHo Kim <hao@smartmind.team>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
langchain-openai = "^0.2.12"
langchain-postgres = "^0.0.12"
langchain-core = "^0.3.25"
langchain = "^0.3.12"
fastapi = "0.111.1"
uvicorn = "0.25"
openai = "^1.57.4"
pglast = "^7.1"
pydantic-settings = "^2.7.0"
redis = "^5.2.1"
tabulate = "^0.9.0"
thanosql = "^1.0.11"
chainlit = "1.3.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
